Finalize ThreeTable SBRG for merging
Old SBRG files moved to backup.
Comments added.
